% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{action_idx_df}
\alias{action_idx_df}
\title{Info about the actions in the HMDP model under consideration.}
\usage{
action_idx_df(prefix = "", file = "actionIdx.bin", labels = "actionIdxLbl.bin")
}
\arguments{
\item{prefix}{A character string with the prefix added to til file(s).}

\item{file}{The HMDP binary file containing the description under consideration.}

\item{labels}{The HMDP binary file containing the labels under consideration.}
}
\value{
A data frame with the same columns as in
\code{action_idx_mat} plus another column containing the labels.
}
\description{
Info about the actions in the HMDP model under consideration.
}
\keyword{internal}
