% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{action_info}
\alias{action_info}
\title{Info about the actions in the HMDP model under consideration.}
\usage{
action_info(
  prefix = "",
  file = "actionIdx.bin",
  weight_file = "actionWeight.bin",
  trans_pr_file = "transProb.bin",
  labels = "actionIdxLbl.bin"
)
}
\arguments{
\item{prefix}{A character string with the prefix added to til file(s).}

\item{file}{The HMDP binary file containing the description under consideration.}

\item{weight_file}{The HMDP binary file containing the action costs.}

\item{trans_pr_file}{The HMDP binary file containing the transition probabilities.}

\item{labels}{The HMDP binary file containing the labels under consideration.}
}
\value{
A matrix with columns from \code{action_idx_mat},
\code{actionCostMat} and \code{trans_prob_mat} if labels is NULL. If labels
not are NULL then a data frame are returned with a label column too.
}
\description{
Info about the actions in the HMDP model under consideration.
}
\keyword{internal}
