% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{action_weight_mat}
\alias{action_weight_mat}
\title{Info about the weights of the actions in the HMDP model under consideration.}
\usage{
action_weight_mat(
  prefix = "",
  file = "actionWeight.bin",
  labels = "actionWeightLbl.bin"
)
}
\arguments{
\item{prefix}{A character string with the prefix added to til file(s).}

\item{file}{The HMDP binary file containing the description under consideration.}

\item{labels}{The HMDP binary file containing the labels under consideration.}
}
\value{
A matrix with columns (\code{aId}, ...) where
\code{aId} is the action row id and ... are the weights of the action.
}
\description{
Info about the weights of the actions in the HMDP model under consideration.
}
\keyword{internal}
