% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binarymodel}
\alias{binarymodel}
\title{Try to solve MDP with binary exact model}
\usage{
binarymodel(distances, m, MAX_TIME = 10, THREADS = 4L,
  verbose = FALSE)
}
\arguments{
\item{\code{distanceMatrix}}{Square and symmetric distance matrix}

\item{\code{m}}{Tour size}

\item{\code{MAX_TIME}}{Its time to stop folks.}

\item{\code{THREADS}}{Number of threads to be used}

\item{\code{verbose}}{true or false}
}
\value{
A baby
}
\description{
Try to solve MDP with binary exact model
}
\examples{
binarymodel(...)
}
