% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{binInfoActions}
\alias{binInfoActions}
\title{Info about the actions in the HMDP model under consideration.}
\usage{
binInfoActions(
  prefix = "",
  labels = TRUE,
  fileA = "actionIdx.bin",
  filePr = "transProb.bin",
  fileW = "actionWeight.bin",
  fileLabelW = "actionWeightLbl.bin",
  fileLabelA = "actionIdxLbl.bin"
)
}
\arguments{
\item{prefix}{A character string with the prefix added to til binary files.}

\item{labels}{Should labels be extracted.}

\item{fileA}{The binary file containing the description of actions.}

\item{filePr}{The binary file containing the description of transition probabilities.}

\item{fileW}{The binary file containing the description of weights.}

\item{fileLabelW}{The binary file containing the weight labels.}

\item{fileLabelA}{The binary file containing the action labels.}
}
\value{
A data frame with the information. Scope string contain the scope of the transitions and
can be 4 values: 2 - A transition to a child process (stage zero in the child process), 1 - A
transition to next stage in the current process, 0 - A transition to the next stage in the
father process. Finally, if scope = 3 then a transition to the state with sId = idx is
considered. The index string denote the index (id is scope = 3) of the state at the next stage.
}
\description{
Info about the actions in the HMDP model under consideration.
}
\note{
The model don't have to be loaded, i.e only read the binary files. The state id (sId) will
not be the same as in the loaded model!
}
\author{
Lars Relund \email{lars@relund.dk}
}
