\name{calcSteadyStatePr}
\alias{calcSteadyStatePr}
\title{Calculate the steady state transition probabilities for the founder process (level 0).}
\usage{calcSteadyStatePr(mdp)}
\description{Calculate the steady state transition probabilities for the founder process (level 0).}
\details{Assume that we consider an ergodic/irreducible time-homogeneous Markov chain specified using a policy in the MDP.}
\value{A vector stady state probabilities for all the states.}
\author{Lars Relund \email{lars@relund.dk}}
\arguments{\item{mdp}{The MDP loaded using \link{loadMDP}.}}
