% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{checkWIdx}
\alias{checkWIdx}
\alias{.check_w_idx}
\title{Internal function. Check if the index of the weight is okay. Should not be used except you know what you are doing}
\usage{
.check_w_idx(i_w, w_lth)
}
\arguments{
\item{i_w}{Index of the weight we want to optimize.}

\item{w_lth}{Number of weights in the model.}
}
\value{
Nothing.
}
\description{
Internal function. Check if the index of the weight is okay. Should not be used except you know what you are doing
}
\keyword{internal}
