% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{compute_gene_score}
\alias{compute_gene_score}
\title{Compute gene score
Computes gene scores for each gene within each class and perturbation freq}
\usage{
compute_gene_score(zscore, pdata, control_lab, score_type)
}
\description{
Compute gene score
Computes gene scores for each gene within each class and perturbation freq
}
