% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{compute_perturbed_genes}
\alias{compute_perturbed_genes}
\title{Compute perturbed genes
Find the top fraction of genes that are more perturbed in test versus controls}
\usage{
compute_perturbed_genes(gmdp_results, control_lab, fraction_genes)
}
\description{
Compute perturbed genes
Find the top fraction of genes that are more perturbed in test versus controls
}
