% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{compute_sample_scores}
\alias{compute_sample_scores}
\title{Compute sample scores for each pathway}
\usage{
compute_sample_scores(zscore, perturbed_genes, control_samples, test_samples,
  pathways, pdata)
}
\description{
Compute sample scores for each pathway
}
