% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_binary_to_hmp}
\alias{convert_binary_to_hmp}
\title{Convert a HMDP model stored in binary format to a \code{hmp} (XML) file.
The function simply parse the binary files and create \code{hmp} files using
the \code{\link[=hmp_mdp_writer]{hmp_mdp_writer()}}.}
\usage{
convert_binary_to_hmp(
  prefix = "",
  bin_names = c("stateIdx.bin", "stateIdxLbl.bin", "actionIdx.bin", "actionIdxLbl.bin",
    "actionWeight.bin", "actionWeightLbl.bin", "transProb.bin"),
  out = paste0(prefix, "converted.hmp"),
  duration = 1,
  get_log = TRUE
)
}
\arguments{
\item{prefix}{A character string with the prefix which will be added to the binary files.}

\item{bin_names}{A character vector of length 7 giving the names of the binary files storing the model.}

\item{out}{The name of the HMP file (e.g. \code{r.hmp}).}

\item{duration}{Weight number storing the duration (NULL if none).}

\item{get_log}{Output log text.}
}
\value{
NULL (invisible).
}
\description{
Convert a HMDP model stored in binary format to a \code{hmp} (XML) file.
The function simply parse the binary files and create \code{hmp} files using
the \code{\link[=hmp_mdp_writer]{hmp_mdp_writer()}}.
}
\note{
Note all indexes are starting from zero (C/C++ style).
}
\examples{
## Set working dir
fDir <- system.file("models", package = "MDP2")
wd <- setwd(tempdir())
## Convert the machine example to a hmp file
prefix1 <- paste0(fDir,"/machine1_")
get_bin_info_states(prefix1)
convert_binary_to_hmp(prefix1, duration = NULL, out = "machine1_converted.hmp")
# have a look at the hmp file
cat(readr::read_file("machine1_converted.hmp"))

## Convert the machine example hmp file to binary files
convert_hmp_to_binary(file = paste0(fDir,"/machine1.hmp"), prefix = "machine_cov_")
get_bin_info_states(prefix = "machine_cov_")
## Convert the machine example with a single dummy node to a hmp file
#convert_binary_to_hmp("machine2_")  # error since using scope = 3 not supported in hmp files

## Reset working dir
setwd(wd)
}
\seealso{
\code{\link[=convert_hmp_to_binary]{convert_hmp_to_binary()}}.
}
