% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_hmp_to_binary}
\alias{convert_hmp_to_binary}
\title{Convert a HMDP model stored in a hmp (xml) file to binary file format.}
\usage{
convert_hmp_to_binary(file, prefix = "", get_log = TRUE)
}
\arguments{
\item{file}{The name of the HMP file (e.g. \code{r.hmp}).}

\item{prefix}{A character string with the prefix which will be added to the binary files.}

\item{get_log}{Output log text.}
}
\value{
NULL (invisible).
}
\description{
The function simply parse the hmp file and create binary files using
the \code{\link[=binary_mdp_writer]{binary_mdp_writer()}}.
}
\note{
Note all indexes are starting from zero (C/C++ style).
}
\examples{
## Set working dir
fDir <- system.file("models", package = "MDP2")
wd <- setwd(tempdir())
## Convert the machine example to a hmp file
prefix1 <- paste0(fDir,"/machine1_")
get_bin_info_states(prefix1)
convert_binary_to_hmp(prefix1, duration = NULL, out = "machine1_converted.hmp")
# have a look at the hmp file
cat(readr::read_file("machine1_converted.hmp"))

## Convert the machine example hmp file to binary files
convert_hmp_to_binary(file = paste0(fDir,"/machine1.hmp"), prefix = "machine_cov_")
get_bin_info_states(prefix = "machine_cov_")
## Convert the machine example with a single dummy node to a hmp file
#convert_binary_to_hmp("machine2_")  # error since using scope = 3 not supported in hmp files

## Reset working dir
setwd(wd)
}
\seealso{
\code{\link[=binary_mdp_writer]{binary_mdp_writer()}}.
}
