% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_prep.R
\name{def_prep}
\alias{def_prep}
\title{Operational data preparation (optimized)}
\usage{
def_prep(folder, ano_inicial, ano_final, mes_inicial, mes_final)
}
\arguments{
\item{folder}{Path to shapefiles folder}

\item{ano_inicial}{Initial year}

\item{ano_final}{Final year}

\item{mes_inicial}{Initial month}

\item{mes_final}{Final month}
}
\value{
data.frame with aggregated variables
}
\description{
Operational data preparation (optimized)
}
