% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_prep_prodes.R
\name{def_prep_prodes}
\alias{def_prep_prodes}
\title{Prepare dataset for monthly prioritization}
\usage{
def_prep_prodes(folder, ano_inicial, ano_final, mes_inicial, mes_final, annual)
}
\arguments{
\item{folder}{Path to data folder}

\item{ano_inicial}{Start year}

\item{ano_final}{End year}

\item{mes_inicial}{Start month}

\item{mes_final}{End month}

\item{annual}{Logical, whether to run PRODES or DETER}
}
\value{
Data frame with aggregated results
}
\description{
Aggregates deforestation, fire, and distance variables into 25km grid.
}
