% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadMDP.R
\name{deleteHMDP}
\alias{.deleteHMDP}
\alias{deleteHMDP}
\title{Internal function. Remove the HMDP from memory. Should not be used except you know what you are doing}
\usage{
.deleteHMDP(p)
}
\arguments{
\item{p}{External pointer to the model.}
}
\value{
Nothing.
}
\description{
Internal function. Remove the HMDP from memory. Should not be used except you know what you are doing
}
\author{
Lars Relund \email{lars@relund.dk}
}

