% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmamdp}
\alias{dmamdp}
\title{Hao's Diversification-driven Memetic Algorith for Maximum Diversity Problem}
\usage{
dmamdp(distanceMatrix, tourSize, populationSize, maxTime,
  lostMaxIterations, p)
}
\arguments{
\item{\code{distanceMatrix}}{Symmetric matrix.}

\item{\code{tourSize}}{Subset size of nodes.}

\item{\code{populationSize}}{Number of individual in population.}

\item{\code{lostMaxIterations}}{for the tabu search.}

\item{\code{maxTime}}{Time limit for execution.}

\item{\code{p}}{Probability of get laid. Otherwise go fish in random pool.}
}
\value{
A better man.
}
\description{
Hao's Diversification-driven Memetic Algorith for Maximum Diversity Problem
}
\details{
Get over it!
}
