% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doBackboneCrossOver}
\alias{doBackboneCrossOver}
\title{Backbone Cross Over between S_a and S_b}
\usage{
doBackboneCrossOver(S_a, S_b, distanceMatrix)
}
\arguments{
\item{\code{S_a}}{Parent A}

\item{\code{S_b}}{Parent B}

\item{\code{distanceMatrix}}{Square and symmetric distance matrix}
}
\value{
A baby
}
\description{
Backbone Cross Over between S_a and S_b
}
\examples{
doBackboneCrossOver()
}
