% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doTabuSearchParallel}
\alias{doTabuSearchParallel}
\title{Execute a Tabu Search in the neighborhood of Solution S (Max Iterations)}
\usage{
doTabuSearchParallel(S, distanceMatrix, alpha = 15L, rhoOver2 = 4,
  maxIterations = 1000L)
}
\arguments{
\item{\code{S}}{initial solution}

\item{\code{distanceMatrix}}{Square and symmetric distance matrix}
}
\value{
Best solution of local tabu search
}
\description{
Execute a Tabu Search in the neighborhood of Solution S (Max Iterations)
}
\examples{
dotabuSearch()
}
