% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_drive_data}
\alias{download_drive_data}
\title{Download shapefiles and tables from Google Drive}
\usage{
download_drive_data(folder_id, folder_id_tabelas, dest_folder = "data/raw")
}
\arguments{
\item{folder_id}{Google Drive folder ID for shapefiles}

\item{folder_id_tabelas}{Google Drive folder ID for tables}

\item{dest_folder}{Local destination folder}
}
\value{
List with local paths
}
\description{
Downloads all shapefiles and tables from public Google Drive folders.
}
