% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Expression data example}
\format{A data frame with 13838 rows and 40 variables:}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE17156}
}
\usage{
example_data
}
\description{
\describe{
  \item{rownames}{HGNC gene names}
  \item{colnames}{sample expression data}
  ...
}
}
\details{
Author expression data for GEO dataset GSE17156 of transcriptome blood samples from patients that were inoculated with the RSV virus that 
has been altered by collapsing for HGNC gene symbols.
}
\keyword{datasets}
