% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pheno}
\alias{example_pheno}
\title{Phenotypic data example}
\format{A data frame with 40 rows and 2 variables:
\describe{
  \item{Sample}{GSM identified}
  \item{Class}{Symtpomatic state}
  ...
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE17156}
}
\usage{
example_pheno
}
\description{
Subset of the annotation data for GEO dataset GSE17156, using only patients that have been inoculated with the RSV virus
}
\keyword{datasets}
