% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{find_deter_data}
\alias{find_deter_data}
\title{Find deter_data folder automatically}
\usage{
find_deter_data(base_path = NULL)
}
\arguments{
\item{base_path}{Optional user-provided path}
}
\value{
Path to deter_data/base/shp
}
\description{
Searches for deter_data folder in common locations.
}
