% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findBestPopFitness}
\alias{findBestPopFitness}
\title{Best Fitness and solution index of population}
\usage{
findBestPopFitness(Population, distanceMatrix)
}
\arguments{
\item{\code{Population}}{to fit.}

\item{\code{distanceMatrix}}{Square and symmetric distance matrix}
}
\value{
A baby
}
\description{
Best Fitness and solution index of population
}
\examples{
findBestPopFitness()
}
