% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache_dir.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get cache directory}
\usage{
get_cache_dir()
}
\value{
Character string containing the cache path.
}
\description{
Returns the local cache directory used by pdm.
}
