% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{get_rpo}
\alias{get_rpo}
\title{Calculate the retention pay-off (RPO) or opportunity cost for some states.}
\usage{
get_rpo(
  mdp,
  w,
  i_a,
  s_id = ifelse(mdp$time_horizon >= Inf, mdp$founder_states_last + 1,
    1):ifelse(mdp$time_horizon >= Inf, mdp$states + mdp$founder_states_last, mdp$states)
    - 1,
  criterion = "expected",
  dur = "",
  rate = 0,
  rate_base = 1,
  discount_factor = NULL,
  g = 0,
  objective = c("max", "min"),
  discount_method = "continuous",
  state_str = TRUE
)
}
\arguments{
\item{mdp}{The MDP loaded using \code{\link[=load_mdp]{load_mdp()}}.}

\item{w}{The label of the weight we calculate RPO for.}

\item{i_a}{The action index we calculate the RPO with respect to (same size as \code{s_id}).}

\item{s_id}{Vector of id's of the states we want to retrieve.}

\item{criterion}{The Bellman operator shortcut. If \code{expected} use expected weights, if \code{discount} use discounted expected weights, if \code{average} use average expected weights.}

\item{dur}{The label of the duration/time such that discount rates can be calculated.}

\item{rate}{The interest rate.}

\item{rate_base}{The time-horizon the rate is valid over.}

\item{discount_factor}{The discount rate for one time unit. If specified \code{rate} and \code{rate_base} are not used to calculate the discount rate.}

\item{g}{The optimal gain (g) calculated (used if \code{criterion = "average"}).}

\item{objective}{Optimize by maximizing (\code{"max"}) or minimizing (\code{"min"}) the Bellman value.}

\item{discount_method}{Either 'continuous' or 'discrete', corresponding to discount factor \code{exp(-rate/rate_base)} or \code{1/(1 + rate/rate_base)}, respectively. Only used if \code{discount_factor} is \code{NULL}.}

\item{state_str}{Output the state string.}
}
\value{
The RPO (matrix/data frame).
}
\description{
The RPO is defined as the difference between
the weight of the state when using action \code{i_a} and the maximum
weight of the node when using another predecessor different from \code{i_a}.
}
