% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{get_steady_state_pr}
\alias{get_steady_state_pr}
\title{Calculate the steady state transition probabilities for the founder process (level 0).}
\usage{
get_steady_state_pr(mdp, get_log = FALSE)
}
\arguments{
\item{mdp}{The MDP loaded using \code{\link[=load_mdp]{load_mdp()}}.}

\item{get_log}{Output log text.}
}
\value{
A vector with steady state probabilities for all the states at the founder level.
}
\description{
Assume that we consider an ergodic/irreducible time-homogeneous Markov chain specified using a policy in the MDP.
}
