% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{get_w_idx}
\alias{get_w_idx}
\title{Return the index of a weight in the model. Note that index always start from zero (C++ style), i.e. the first weight, the first state at a stage etc has index 0.}
\usage{
get_w_idx(mdp, w_lbl)
}
\arguments{
\item{mdp}{The MDP loaded using \code{\link[=load_mdp]{load_mdp()}}.}

\item{w_lbl}{The label/string of the weight.}
}
\value{
The index (integer).
}
\description{
Return the index of a weight in the model. Note that index always start from zero (C++ style), i.e. the first weight, the first state at a stage etc has index 0.
}
