% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAverageDistanceToPopulation}
\alias{getAverageDistanceToPopulation}
\title{Average distance between a individual and its population;}
\usage{
getAverageDistanceToPopulation(S0, Population)
}
\arguments{
\item{\code{S0}}{Individual.}

\item{\code{Population}}{Target Population.}
}
\value{
A int, min distance between a indivitual and a population.
}
\description{
Average distance between a individual and its population;
}
\details{
Average distance between a individual and other elements in population: mean(m - sum(S0XS), S in Population, S not i);
}
