% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAverageDistanceToPopulationByIndex}
\alias{getAverageDistanceToPopulationByIndex}
\title{Average distance between a individual and its population (by index);}
\usage{
getAverageDistanceToPopulationByIndex(index, Population)
}
\arguments{
\item{\code{S0}}{Individual.}
}
\value{
A int, min distance between a indivitual and a population.
}
\description{
Average distance between a individual and its population (by index);
}
\details{
Average distance between a individual and other elements in population: mean(m - sum(S0XS), S in Population, S not i);
}
