% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSolutionsDistance}
\alias{getSolutionsDistance}
\title{Distance between to individuals;}
\usage{
getSolutionsDistance(S0, S1)
}
\arguments{
\item{\code{S0}}{Individual.}

\item{\code{S1}}{Individual.}
}
\value{
A int, number of moves to go from one individual to the other.
}
\description{
Distance between to individuals;
}
\details{
Distance between to individuals: m - sum(S0XS1);
}
