% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSolutionToPopulationDistance}
\alias{getSolutionToPopulationDistance}
\title{Distance between a individual and a population;}
\usage{
getSolutionToPopulationDistance(S0, Population)
}
\arguments{
\item{\code{S0}}{Individual.}
}
\value{
A int, min distance between a indivitual and a population.
}
\description{
Distance between a individual and a population;
}
\details{
Distance between to individuals in population: min(m - sum(S0XS), S in Population);
}
