% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSolutionToPopulationDistanceByIndex}
\alias{getSolutionToPopulationDistanceByIndex}
\title{Distance between a individual and a population by index;}
\usage{
getSolutionToPopulationDistanceByIndex(index, Population)
}
\arguments{
\item{\code{S0}}{Individual.}

\item{\code{Population}}{Target Population.}
}
\value{
A int, min distance between a indivitual and a population.
}
\description{
Distance between a individual and a population by index;
}
\details{
Distance between to individuals in population: min(m - sum(S0XS), S in Population);
}
