% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{getSteadyStatePr}
\alias{getSteadyStatePr}
\title{Calculate the steady state transition probabilities for the founder process (level 0).}
\usage{
getSteadyStatePr(mdp, getLog = FALSE)
}
\arguments{
\item{mdp}{The MDP loaded using \link{loadMDP}.}

\item{getLog}{Output log text.}
}
\value{
A vector with stady state probabilities for all the states at the founder level.
}
\description{
Assume that we consider an ergodic/irreducible time-homogeneous Markov chain specified using a policy in the MDP.
}
\author{
Lars Relund \email{lars@relund.dk}
}
