\name{getTransPr}
\alias{getTransPr}
\title{Get the transition probability matrix P for the founder process (level 0).}
\usage{getTransPr(mdp)}
\description{Get the transition probability matrix P for the founder process (level 0).}
\value{The state probability matrix.}
\author{Lars Relund \email{lars@relund.dk}}
\arguments{\item{mdp}{The MDP loaded using \link{loadMDP}.}}
