% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializeOPB}
\alias{initializeOPB}
\title{Opposition Based Population initialization}
\usage{
initializeOPB(distanceMatrix, tourSize, populationSize = 10L,
  maxIterations = 100L, multiplier = 2L)
}
\arguments{
\item{\code{TourSize}}{Well..}

\item{\code{Distances}}{Distance matrix}

\item{\code{PopulationSize}}{That's it.}
}
\value{
A matrix where each column is a individual in the population
}
\description{
Opposition Based Population initialization
}
\details{
Get the initial populatio
}
