% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializePool}
\alias{initializePool}
\title{Pool Population initialization}
\usage{
initializePool(distanceMatrix, tourSize, populationSize = 10L,
  maxIterations = 100L, multiplier = 3L)
}
\arguments{
\item{\code{tourSize}}{Ok. Stop it.}

\item{\code{Distances}}{Distance matrix}

\item{\code{populationSize}}{You know!}
}
\value{
A matrix where each column is a individual in the population
}
\description{
Pool Population initialization
}
\details{
Get the initial populatio
}
