% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializePoolML}
\alias{initializePoolML}
\title{Pool Population initialization (Maximum Lost)}
\usage{
initializePoolML(distanceMatrix, tourSize, populationSize = 10L,
  lostMaxIterations = 100L, multiplier = 1L)
}
\arguments{
\item{\code{TourSize}}{Thats it.}

\item{\code{Distances}}{Distance matrix}

\item{\code{populationSize}}{Same old guy}
}
\value{
A matrix where each column is a individual in the population
}
\description{
Pool Population initialization (Maximum Lost)
}
\details{
Get the initial populatio
}
