% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{load_ext_csv}
\alias{load_ext_csv}
\title{Load external shapefile}
\usage{
load_ext_csv(name, package = "pdm")
}
\arguments{
\item{name}{Name of the spatial file.}

\item{package}{Package containing the file.}
}
\value{
An sf object.
}
\description{
Reads a spatial file from inst/extdata.
}
