% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mamdp}
\alias{mamdp}
\title{Hao's Hybrid Metaheuristic method for the Maximum Diversity Problem}
\usage{
mamdp(distanceMatrix, tourSize, populationSize, maxIterations,
  maxTime = 60, multiplier = 3L)
}
\arguments{
\item{\code{distanceMatrix}}{Symmetric matrix.}

\item{\code{tourSize}}{Subset size of nodes.}

\item{\code{populationSize}}{Number of individual in population.}

\item{\code{maxIterations}}{for the tabu search.}
}
\value{
A better person.
}
\description{
Hao's Hybrid Metaheuristic method for the Maximum Diversity Problem
}
\details{
Get over it!
}
