\name{mdp-package}
\alias{mdp-package}
\alias{mdp}
\docType{package}
\title{
\packageTitle{mdp}
}
\description{
\packageDescription{mdp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mdp}
\packageIndices{mdp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mdp}

Maintainer: \packageMaintainer{mdp}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
