% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obma}
\alias{obma}
\title{Hao's Opposition-based Memetic memetic search for the Maximum Diversity Problem}
\usage{
obma(distanceMatrix, tourSize, populationSize, maxIterations)
}
\arguments{
\item{\code{distanceMatrix}}{Symmetric matrix.}

\item{\code{tourSize}}{Subset size of nodes.}

\item{\code{populationSize}}{Number of individual in population.}

\item{\code{maxIterations}}{for the tabu search.}
}
\value{
A better man.
}
\description{
Hao's Opposition-based Memetic memetic search for the Maximum Diversity Problem
}
\details{
Get over it!
}
