% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{plot.smdp}
\alias{plot.smdp}
\title{Plot sMDP scores}
\usage{
\method{plot}{smdp}(sMDP.plot, filename = "", path = "", title.graph = "")
}
\arguments{
\item{filename}{filename}

\item{sMDP.data}{a data.frame containing sMDP information for a geneset, with columns "Sample", "sMDP" and "Class"}
}
\description{
Plot sMDP scores
}
