% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_pipeline.R
\name{prep_pipeline}
\alias{prep_pipeline}
\title{Full pipeline: data prep + model}
\usage{
prep_pipeline(
  folder_id = NULL,
  folder_id_tabelas = NULL,
  base_path = NULL,
  ano_inicial,
  ano_final,
  mes_inicial,
  mes_final,
  run_model = TRUE
)
}
\arguments{
\item{folder_id}{Google Drive folder ID (optional)}

\item{folder_id_tabelas}{Google Drive folder ID (optional)}

\item{base_path}{Optional local path}

\item{ano_inicial}{Initial year}

\item{ano_final}{Final year}

\item{mes_inicial}{Initial month}

\item{mes_final}{Final month}

\item{run_model}{Should run def_prio? (default = TRUE)}
}
\description{
Full pipeline: data prep + model
}
