% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{print_pathways}
\alias{print_pathways}
\title{print pathways
summary plot for pathways and sample score plot of best gene set}
\usage{
print_pathways(sample_results, path, file_name)
}
\description{
print pathways
summary plot for pathways and sample score plot of best gene set
}
