% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_hmdp.R
\name{random_hmdp}
\alias{random_hmdp}
\title{Generate a "random" HMDP stored in a set of binary files.}
\usage{
random_hmdp(
  prefix = "",
  levels = 3,
  time_horizon = c(Inf, 3, 4),
  states = c(2, 4, 5),
  actions = c(1, 2),
  child_process_pr = 0.5,
  external_process_pr = 0,
  rewards = c(0, 100),
  durations = c(1, 10),
  reward_name = "Reward",
  duration_name = "Duration"
)
}
\arguments{
\item{prefix}{A character string with the prefix added to the file(s).}

\item{levels}{Maximum number of levels. Set \code{child_process_pr = 1} if want exact this number of levels.}

\item{time_horizon}{The time horizon for each level (vector). For the founder the time-horizon can be Inf.}

\item{states}{Number of states at each stage at a given level (vector of length levels)}

\item{actions}{Min and max number of actions at a state.}

\item{child_process_pr}{Probability of creating a child process when define action.}

\item{external_process_pr}{Probability of creating an external process given that we create a child process. Only works if levels>2 and and currently does not generate external processes which include external processes.}

\item{rewards}{Min and max reward used.}

\item{durations}{Min and max duration used.}

\item{reward_name}{Weight name used for reward.}

\item{duration_name}{Weight name used for duration.}
}
\value{
The file prefix (character).
}
\description{
Generate a "random" HMDP stored in a set of binary files.
}
