% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/randomHMDP.R
\name{randomHMDP}
\alias{randomHMDP}
\title{Generate a "random" HMDP stored in a set of binary files.}
\usage{
randomHMDP(prefix = "", levels = 3, timeHorizon = c(Inf, 3, 4),
  states = c(2, 4, 5), actions = c(1, 2), childProcessPr = 0.5,
  rewards = c(0, 100), durations = c(1, 10))
}
\arguments{
\item{prefix}{A character string with the prefix added to til file(s).}

\item{levels}{Number of levels.}

\item{timeHorizon}{The time horizon for each level (vector). For the founder the timehorizon can be Inf.}

\item{states}{Number of states at each stage at a given level (vector of length levels)}

\item{actions}{Min and max number of actions at a state.}

\item{childProcessPr}{Probability of creating a child process when define action.}

\item{rewards}{Min and max reward used.}

\item{durations}{Min and max duration used.}
}
\description{
Generate a "random" HMDP stored in a set of binary files.
}
\author{
Lars Relund \email{lars@relund.dk}
}

