% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_all_shapefiles}
\alias{read_all_shapefiles}
\title{Read all shapefiles from a folder}
\usage{
read_all_shapefiles(folder)
}
\arguments{
\item{folder}{Path to folder containing shapefiles}
}
\value{
Named list of sf objects
}
\description{
Loads all .shp files in a directory into a list of sf objects.
}
