% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{read_gmt}
\alias{read_gmt}
\title{Read gmt}
\usage{
read_gmt(fname)
}
\arguments{
\item{fname}{the path to your gene matrix transposed file format (*.gmt) file}
}
\description{
File to read a gmt file and return as a list
}
