% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{run_policy_ite_ave}
\alias{run_policy_ite_ave}
\title{Perform policy iteration using the average expected-weight Bellman operator on the MDP.}
\usage{
run_policy_ite_ave(
  mdp,
  w,
  dur,
  max_ite = 100,
  objective = c("max", "min"),
  get_log = TRUE
)
}
\arguments{
\item{mdp}{The MDP loaded using \code{\link[=load_mdp]{load_mdp()}}.}

\item{w}{The label of the weight we optimize.}

\item{dur}{The label of the duration/time such that discount rates can be calculated.}

\item{max_ite}{Max number of iterations. If the model does not satisfy the unichain assumption the algorithm may loop.}

\item{objective}{Optimize by maximizing (\code{"max"}) or minimizing (\code{"min"}) the Bellman value.}

\item{get_log}{Output the log messages.}
}
\value{
The optimal gain (g) calculated.
}
\description{
The policy can afterwards be received using functions \code{get_policy} and \code{get_policy_w}.
}
\seealso{
\code{\link[=get_policy]{get_policy()}}.
}
