% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{run_policy_ite_discount}
\alias{run_policy_ite_discount}
\title{Perform policy iteration using the discounted expected-weight Bellman operator on the MDP.}
\usage{
run_policy_ite_discount(
  mdp,
  w,
  dur,
  rate = 0,
  rate_base = 1,
  discount_factor = NULL,
  max_ite = 100,
  discount_method = "continuous",
  objective = c("max", "min"),
  get_log = TRUE
)
}
\arguments{
\item{mdp}{The MDP loaded using \code{\link[=load_mdp]{load_mdp()}}.}

\item{w}{The label of the weight we optimize.}

\item{dur}{The label of the duration/time such that discount rates can be calculated.}

\item{rate}{The interest rate.}

\item{rate_base}{The time-horizon the rate is valid over.}

\item{discount_factor}{The discount rate for one time unit. If specified \code{rate} and \code{rate_base} are not used to calculate the discount rate.}

\item{max_ite}{Max number of iterations. If the model does not satisfy the unichain assumption the algorithm may loop.}

\item{discount_method}{Either 'continuous' or 'discrete', corresponding to discount factor \code{exp(-rate/rate_base)} or \code{1/(1 + rate/rate_base)}, respectively. Only used if \code{discount_factor} is \code{NULL}.}

\item{objective}{Optimize by maximizing (\code{"max"}) or minimizing (\code{"min"}) the Bellman value.}

\item{get_log}{Output the log messages.}
}
\value{
Nothing.
}
\description{
The policy can afterwards be received using functions \code{get_policy} and \code{get_policy_w}.
}
\seealso{
\code{\link[=get_policy]{get_policy()}}.
}
