% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample score results}
\format{A data frame with 40 rows and 3 variables:
\describe{
  \item{Sample}{GSM identified}
  \item{Score}{Sample score}
  \item{Class}{Symtpomatic state}
  ...
}}
\usage{
sample_data
}
\description{
Resultant sample scores when the mdp is applied to example_data and example_pheno
}
\keyword{datasets}
