% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{save_mdp}
\alias{save_mdp}
\title{Save the MDP to binary files}
\usage{
save_mdp(mdp, prefix = "", get_log = TRUE)
}
\arguments{
\item{mdp}{The MDP loaded using \code{\link[=load_mdp]{load_mdp()}}.}

\item{prefix}{A character string with the prefix added to \code{bin_names}. Used to identify a specific model.}

\item{get_log}{Output the log as a message.}
}
\value{
???
}
\description{
Currently do not save external files.
}
