% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{saveMDP}
\alias{saveMDP}
\title{Save the MDP to binary files}
\usage{
saveMDP(mdp, prefix = "", getLog = TRUE)
}
\arguments{
\item{mdp}{The MDP loaded using \link{loadMDP}.}

\item{prefix}{A character string with the prefix added to \code{binNames}. Used to identify a specific model.}

\item{getLog}{Output the log as a message.}
}
\value{
The rpo (matrix/data frame).
}
\description{
Currently do not save external files.
}
\author{
Lars Relund \email{lars@relund.dk}
}

