% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{set_action_weight}
\alias{set_action_weight}
\title{Set an action weight in a loaded MDP}
\usage{
set_action_weight(mdp, weight, s_id, a_idx, weight_name)
}
\arguments{
\item{mdp}{The MDP loaded using \code{\link[=load_mdp]{load_mdp()}} or created using
\code{\link[=memory_mdp_writer]{memory_mdp_writer()}}.}

\item{weight}{A finite numeric scalar with the new action weight.}

\item{s_id}{Zero-based state index.}

\item{a_idx}{Zero-based action index within the state.}

\item{weight_name}{The name of an action-level weight.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Modifies one action-level weight in the in-memory model. Use \code{\link[=save_mdp]{save_mdp()}} if
the modified model should be written to binary files.
}
