\name{setStateWeight}
\alias{setStateWeight}
\title{Set the weight of a state.}
\usage{setStateWeight(mdp, w, sId, iW)}
\description{Set the weight of a state.}
\value{Nothing.}
\author{Lars Relund \email{lars@relund.dk}}
\arguments{\item{mdp}{The MDP loaded using \link{loadMDP}.}
\item{w}{The weight.}
\item{sId}{The state id of the state.}
\item{iW}{Weight index where set the weight.}}
