% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{smdp_plot}
\alias{smdp_plot}
\title{Plot sMDP scores}
\usage{
smdp_plot(sample_data, filename = file_name, directory = "", title = "",
  print = T)
}
\arguments{
\item{filename}{filename}

\item{directory}{directory to save file}

\item{sMDP.data}{a data.frame containing sMDP information for a geneset, with columns "Sample", "sMDP" and "Class"}

\item{title.graph}{title name for graph}
}
\description{
Plots the sMDP scores for a given geneset. Takes a data.frame that contains the sMDP scores for one geneset, along with the Sample and Class information. Data.frame must have a sMDP, Sample and Class columns.
}
