% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{state_idx_mat}
\alias{state_idx_mat}
\title{Info about the states in the HMDP model under consideration.}
\usage{
state_idx_mat(prefix = "", file = "stateIdx.bin")
}
\arguments{
\item{prefix}{A character string with the prefix added to til file(s).}

\item{file}{The HMDP binary file containing the description under consideration.}
}
\value{
A matrix with columns \verb{(s_id, n0, s0, a0, ...)} where
\code{s_id} is the state row id, \code{n0} the index of the stage at level 0, \code{s0} the index
of the state and \code{a0} the index of the action. If the HMDP has more
than one level columns index \verb{(d1, s1, a1, ...)} are added.
}
\description{
Info about the states in the HMDP model under consideration.
}
\keyword{internal}
