% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{stateIdxDf}
\alias{stateIdxDf}
\title{Info about the states in the HMDP model under consideration.}
\usage{
stateIdxDf(prefix = "", file = "stateIdx.bin", labels = "stateIdxLbl.bin")
}
\arguments{
\item{prefix}{A character string with the prefix added to the file(s).}

\item{file}{The HMDP binary file containing the description under consideration.}

\item{labels}{The HMDP binary file containing the labels under consideration.}
}
\value{
A data frame with the same columns as in
\code{stateIdxMat} plus another column containing the labels.
}
\description{
Info about the states in the HMDP model under consideration.
}
\keyword{internal}
