% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tour_fitness_binary}
\alias{tour_fitness_binary}
\title{Get fitness from tour}
\usage{
tour_fitness_binary(S, distances)
}
\arguments{
\item{\code{Tour}}{Set of tour's nodes.}

\item{\code{Distances}}{Distance matrix}
}
\value{
A double value representing the chromosome fitness
}
\description{
Get fitness from tour
}
\details{
Get fitness using the tour, m and distance matrix
}
