% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{trans_prob_mat}
\alias{trans_prob_mat}
\title{Info about the transition probabilities in the HMDP model under consideration.}
\usage{
trans_prob_mat(prefix = "", file = "transProb.bin")
}
\arguments{
\item{prefix}{A character string with the prefix added to til file(s).}

\item{file}{The HMDP binary file containing the description under consideration.}
}
\value{
A matrix with columns (\code{aId}, ...) where
\code{aId} is the action row id and ... are the probabilities of the action.
}
\description{
Info about the transition probabilities in the HMDP model under consideration.
}
\keyword{internal}
