% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes_download_package.R
\name{update_deforestation_prodes}
\alias{update_deforestation_prodes}
\title{Amazon biome download update months}
\usage{
update_deforestation_prodes()
}
\description{
Amazon biome download update months
}
