% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fires_download_package.R
\name{update_fires}
\alias{update_fires}
\title{Fires download update months}
\usage{
update_fires(mes_inicial, mes_final, ano_inicial, ano_final)
}
\arguments{
\item{mes_inicial}{Start month}

\item{mes_final}{End month}

\item{ano_inicial}{Start year}

\item{ano_final}{End year}
}
\description{
Fires download update months
}
