% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updatePopulationByRank}
\alias{updatePopulationByRank}
\title{Rank based pool population update}
\usage{
updatePopulationByRank(S, Population, distanceMatrix, beta = 0.6)
}
\arguments{
\item{\code{S}}{population candidate}

\item{\code{Population}}{population to be updated}
}
\value{
New population
}
\description{
Rank based pool population update
}
\examples{
updatePopulationByRank()
}
