% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updatePopulationMAMDP}
\alias{updatePopulationMAMDP}
\title{MAMDP population update}
\usage{
updatePopulationMAMDP(S, Population, distanceMatrix, beta = 0.6)
}
\arguments{
\item{\code{S}}{population candidate}

\item{\code{Population}}{population to be updated}
}
\value{
New population
}
\description{
MAMDP population update
}
\examples{
updatePopulation()
}
