% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{weight_names}
\alias{weight_names}
\title{Names of weights used in actions.}
\usage{
weight_names(prefix = "", labels = "actionWeightLbl.bin")
}
\arguments{
\item{prefix}{A character string with the prefix added to the binary file names.}

\item{labels}{The HMDP binary file containing the weight labels.}
}
\value{
Vector of weight names.
}
\description{
Names of weights used in actions.
}
\keyword{internal}
