% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{actionIdxMat}
\alias{actionIdxMat}
\title{Info about the actions in the HMDP model under consideration.}
\usage{
actionIdxMat(prefix = "", file = "actionIdx.bin")
}
\arguments{
\item{prefix}{A character string with the prefix added to til file(s).}

\item{file}{The HMDP binary file containing the description under consideration.}
}
\value{
A matrix with columns (\code{aId}, ...) where
\code{aId} is the action row id and \code{...} are alternating pairs \verb{(scp, idx)}, one for each
possible transition where \code{scp} is the scope that can be 4 values:
2 - A transition to a child process (stage zero in the child process), 1 - A transition
to next stage in the current process, 0 - A transition to the next stage in the father
process. the idx in the pair denote the index of the state at the stage considered.
Finally, if scope equals 3 then a transition to the state with \code{sId = idx} is considered.
}
\description{
Info about the actions in the HMDP model under consideration.
}
\keyword{internal}
