% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{actionInfo}
\alias{actionInfo}
\title{Info about the actions in the HMDP model under consideration.}
\usage{
actionInfo(
  prefix = "",
  file = "actionIdx.bin",
  weightFile = "actionWeight.bin",
  transPrFile = "transProb.bin",
  labels = "actionIdxLbl.bin"
)
}
\arguments{
\item{prefix}{A character string with the prefix added to til file(s).}

\item{file}{The HMDP binary file containing the description under consideration.}

\item{weightFile}{The HMDP binary file containing the action costs.}

\item{transPrFile}{The HMDP binary file containing the transition probabilities.}

\item{labels}{The HMDP binary file containing the labels under consideration.}
}
\value{
A matrix with columns from \code{actionIdxMat},
\code{actionCostMat} and \code{transProbMat} if labels is NULL. If labels
not are NULL then a data frame are returned with a label column too.
}
\description{
Info about the actions in the HMDP model under consideration.
}
\author{
Lars Relund \email{lars@relund.dk}
}
\keyword{internal}
