% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{calcRPO}
\alias{calcRPO}
\title{Calculate the rentention payoff (RPO) or opportunity cost for some states.}
\usage{
calcRPO(mdp, w, iA, sId = ifelse(mdp$timeHorizon >= Inf, mdp$founderStatesLast
  + 1, 1):ifelse(mdp$timeHorizon >= Inf, mdp$states + mdp$founderStatesLast,
  mdp$states) - 1, criterion = "expected", dur = "", rate = 0,
  rateBase = 1, discountFactor = NULL, g = 0,
  discountMethod = "continuous")
}
\arguments{
\item{mdp}{The MDP loaded using \link{loadMDP}.}

\item{w}{The label of the weight/reward we calculate RPO for.}

\item{iA}{The action index we calculate the RPO with respect to (same size as sId).}

\item{sId}{Vector of id's of the states we want to retrive.}

\item{criterion}{The criterion used. If \code{expected} used expected reward, if \code{discount} used discounted rewards, if \code{average} use average rewards.}

\item{dur}{The label of the duration/time such that discount rates can be calculated.}

\item{rate}{The interest rate.}

\item{rateBase}{The time-horizon the rate is valid over.}

\item{discountFactor}{The discountRate for one time unit. If specified \code{rate} and \code{rateBase} are not used to calculate the discount rate.}

\item{g}{The optimal gain (g) calculated (used if \code{criterion = "average"}).}

\item{discountMethod}{Either 'continuous' or 'discrete', corresponding to discount factor exp(-rate/rateBase) or 1/(1+rate/rateBase), respectively. Only used if \code{discountFactor} is \code{NULL}.}
}
\value{
The rpo (matrix/data frame).
}
\description{
The RPO is defined as the difference between
the weight of the state when using action \code{iA} and the maximum
weight of the node when using another predecessor different from \code{iA}.
}
\author{
Lars Relund \email{lars@relund.dk}
}

