% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadMDP.R
\name{checkWDurIdx}
\alias{.checkWDurIdx}
\alias{checkWDurIdx}
\title{Internal function. Check if the indexes given are okay. Should not be used except you know what you are doing}
\usage{
.checkWDurIdx(iW, iDur, wLth)
}
\arguments{
\item{iW}{Index of the weight we want to optimize.}

\item{iDur}{Index of the duration/time.}

\item{wLth}{Number of weights in the model.}
}
\value{
Nothing.
}
\description{
Internal function. Check if the indexes given are okay. Should not be used except you know what you are doing
}
\author{
Lars Relund \email{lars@relund.dk}
}

