% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{checkWIdx}
\alias{checkWIdx}
\alias{.checkWIdx}
\title{Internal function. Check if the index of the weight is okay. Should not be used except you know what you are doing}
\usage{
.checkWIdx(iW, wLth)
}
\arguments{
\item{iW}{Index of the weight we want to optimize.}

\item{wLth}{Number of weights in the model.}
}
\value{
Nothing.
}
\description{
Internal function. Check if the index of the weight is okay. Should not be used except you know what you are doing
}
\author{
Lars Relund \email{lars@relund.dk}
}
