% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{compute_gene_score}
\alias{compute_gene_score}
\title{Compute gene score
Computes gene scores for each gene within each class and perturbation freq}
\usage{
compute_gene_score(zscore, pdata, control_lab,
  score_type = c("gene_score", "gene_freq"))
}
\arguments{
\item{zscore}{zscore data frame}

\item{pdata}{phenotypic data with Class and Sample columns}

\item{control_lab}{character specifying control class}

\item{score_type}{set to 'gene_score' or 'gene_freq' to compute gene scores 
or frequencies}
}
\value{
data frame of gene scores or gene frequencies
}
\description{
Compute gene score
Computes gene scores for each gene within each class and perturbation freq
}
