% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{compute_perturbed_genes}
\alias{compute_perturbed_genes}
\title{Compute perturbed genes
Find the top fraction of genes that are more perturbed in test versus controls}
\usage{
compute_perturbed_genes(gmdp_results, control_lab, fraction_genes)
}
\arguments{
\item{gmdp_results}{results table of gene scores}

\item{control_lab}{label specificying control class}

\item{fraction_genes}{fraction of top perturbed genes that will make the set of perturbed genes}
}
\description{
Compute perturbed genes
Find the top fraction of genes that are more perturbed in test versus controls
}
