% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{compute_sample_scores}
\alias{compute_sample_scores}
\title{Compute sample scores for each pathway}
\usage{
compute_sample_scores(zscore, perturbed_genes, control_samples, test_samples,
  pathways, pdata)
}
\arguments{
\item{zscore}{zscore data frame}

\item{perturbed_genes}{list of pertured genes}

\item{control_samples}{vector of control sample names}

\item{test_samples}{vector of test sample names}

\item{pathways}{list of pathways}

\item{pdata}{phenotypic data with Sample and Class columns}
}
\description{
Compute sample scores for each pathway
}
