% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{compute_zscore}
\alias{compute_zscore}
\title{Computes the thresholded Z score
Plots the Z score using control samples to compute the average and standard deviation}
\usage{
compute_zscore(data, control_samples, measure, std)
}
\arguments{
\item{data}{gene expression data with gene symbools in rows}

\item{control_samples}{a character vector of control sample names}

\item{measure}{either "mean" or "median". mean uses mean and standard deviation. "median" uses modified z score.}
}
\description{
Computes the thresholded Z score
Plots the Z score using control samples to compute the average and standard deviation
}
