% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_prio.R
\name{def_prio}
\alias{def_prio}
\title{Run prioritization model}
\usage{
def_prio(folder, ano_inicial, ano_final, mes_inicial, mes_final)
}
\arguments{
\item{folder}{Base folder path}

\item{ano_inicial}{Initial year}

\item{ano_final}{Final year}

\item{mes_inicial}{Initial month}

\item{mes_final}{Final month}
}
\description{
Run prioritization model
}
