% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_viz.R
\name{def_viz}
\alias{def_viz}
\title{Visualization and validation of prioritization}
\usage{
def_viz(folder, mes_inicial, mes_final)
}
\arguments{
\item{folder}{Base folder path}

\item{mes_inicial}{Initial month}

\item{mes_final}{Final month}
}
\description{
Visualization and validation of prioritization
}
