% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doTabuSearch}
\alias{doTabuSearch}
\title{Execute a Tabu Search in the neighborhood of Solution S}
\usage{
doTabuSearch(S, distanceMatrix, alpha = 15L, rhoOver2 = 1,
  maxIterations = 100L)
}
\arguments{
\item{S}{initial solution}

\item{distanceMatrix}{}
}
\value{
Best solution of local tabu search
}
\description{
Execute a Tabu Search in the neighborhood of Solution S
}
\examples{
dotabuSearch()
}
