\name{getTransPr}
\alias{getTransPr}
\title{Get the transition probability matrix P for the founder process (level 0).}
\usage{
  getTransPr(mdp)
}
\arguments{
  \item{mdp}{The MDP loaded using \link{loadMDP}.}
}
\value{
  The state probability matrix.
}
\description{
  Get the transition probability matrix P for the founder
  process (level 0).
}
\author{
  Lars Relund \email{lars@relund.dk}
}

