% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{mdp}
\alias{mdp}
\title{Molecular Degree of Perturbation}
\usage{
mdp(data, pdata, control_lab, directory = "", pathways, print = TRUE,
  measure = "mean", std = 2, fraction_genes = 0.25, save_tables = TRUE,
  file_name = "")
}
\arguments{
\item{data}{A data.frame of gene expression data with the gene symbols in the row names}

\item{pdata}{A data.frame of phenodata with a column headed Class and the other headed Sample.}

\item{control_lab}{A character vector of the control class}

\item{directory}{The output directory (optional)}

\item{pathways}{A loaded gmt file in a list format (optional), use read_gmt("gmt.file.location") to load}

\item{print}{Set as default to TRUE if you wish graph pdfs of the geneMDP and sampleMDP values to be printed}

\item{measure}{Set as default to "median", can be changed to "mean". This measure is used in all Z-score calculations.}

\item{std}{Set as default to 2. This controls the standard deviation threshold for the Z-score calculation. Normalised expression values less than "std" will be set to 0.}

\item{save_tables}{Set as default to TRUE. Will save tables of zscore and gene and sample scores.}
}
\value{
A list: [[1]] $zscore [[2]] $gene_scores  [[3]] $gene_freq [[4]] $sample_scores [[5]] perturbed_genes
}
\description{
Based on the Molecular Distance to Health, this function allows the inspection of gene and sample heterogeneity in respect to a control class. When a gmt file is
submitted, the MDP is run on gene subsets. The MDP returns perturbation scores for each gene and each sample.
}
\examples{
basic usage:
mdp(exp,pheno,"healthy_control")
mypathway <- read_gmt("gmt.file.location")
mdp(exp,pheno,"healthy_control",print=TRUE,directory="myexp",pathways=mypathway)
}
