% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{mdp}
\alias{mdp}
\title{Molecular Degree of Perturbation}
\usage{
mdp(data, pdata, control_lab, directory = "", pathways, print = TRUE,
  measure = "median", std = 2)
}
\arguments{
\item{data}{A data.frame of gene expression data with the gene symbols in the rows}

\item{pdata}{A data.frame of phenodata with a column headed Class and the other headed Sample.}

\item{control_lab}{A character vector of the control class}

\item{directory}{The output directory (optional)}

\item{pathways}{The location of a gmt file (optional)}

\item{print}{Set as default to TRUE if you wish graph pdfs of the geneMDP and sampleMDP values to be printed}

\item{measure}{Set as default to "median", can be changed to "mean". This measure is used in all Z-score calculations.}

\item{std}{Set as default to 2. This controls the standard deviation threshold for the Z-score calculation. Normalised expression values less than "std" will be set to 0.}
}
\value{
A list where [[1]] $Zscore contains a table of Z scores, [[2]] $gMDP contains gMDP scores and [[3]] $sMDP are the sampleMDP scores
}
\description{
Based on the Molecular Distance to Health, this function allows the inspection of gene and sample heterogeneity in respect to a control class. When a gmt file is
submitted, the MDP is run on gene subsets. The MDP returns perturbation scores for each gene and each sample.
}
\examples{
mdp(exp,pheno,"healthy_control")
mdp(exp,pheno,"healthy_control",print=TRUE,directory="myexp",pathways="mypathways.gmt")
}
