\name{MDP}
\title{Markov Decision Processes (MDPs) in R}
\description{Create and optimize MDPs or hierarchical MDPs with discrete time steps and state space.}
\details{\tabular{ll}{
Package: \tab MDP\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2009-02-20\cr
License: \tab GPL (>= 2.8)\cr
LazyLoad: \tab yes\cr
}

Create and optimize MDPs or hierarchical MDPs with discrete time steps and state space.}
\alias{MDP-package}
\alias{MDP}
\docType{package}
\author{Lars Relund \email{lars@relund.dk} and Anders Ringgaard Kristensen}
