% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp-package.R
\docType{package}
\name{MDP2-package}
\alias{MDP2}
\alias{MDP2-package}
\title{MDP2: Markov Decision Processes (MDPs) in R}
\description{
Create and optimize (semi) MDPs with discrete time steps and state space. Both hierarchical and ordinary-traditional MDPs can be modelled.
}
\seealso{
\link{loadMDP}.
}
\author{
\strong{Maintainer}: Lars Relund Nielsen \email{lars@relund.dk} (\href{https://orcid.org/0000-0002-4802-3071}{ORCID})

}
\keyword{internal}
