% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{pathway_summary}
\alias{pathway_summary}
\title{print pathways
summary plot for pathways and sample score plot of best gene set}
\usage{
pathway_summary(sample_results, path, file_name, control_samples, control_lab)
}
\description{
print pathways
summary plot for pathways and sample score plot of best gene set
}
