% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mdp.R
\name{policyIteDiscount}
\alias{policyIteDiscount}
\title{Perform policy iteration (discounted reward criterion) on the MDP.}
\usage{
policyIteDiscount(mdp, w, dur, rate = 0.1, rateBase = 1, maxIte = 100)
}
\arguments{
\item{mdp}{The MDP loaded using \link{loadMDP}.}

\item{w}{The label of the weight we optimize.}

\item{dur}{The label of the duration/time such that discount rates can be calculated.}

\item{rate}{The interest rate.}

\item{rateBase}{The time-horizon the rate is valid over.}

\item{maxIte}{Max number of iterations. If the model does not satisfy the unichain assumption the algorithm may loop.}
}
\value{
Nothing.
}
\description{
The policy can afterwards be received using functions \code{getPolicy} and \code{getPolicyW}.
}
\author{
Lars Relund \email{lars@relund.dk}
}
\seealso{
\code{\link{getPolicy}}, \code{\link{getPolicyW}}.
}

