\name{resetActions}
\alias{resetActions}
\title{Reset the actions of a state.}
\usage{resetActions(mdp)}
\description{Reset the actions of a state.}
\value{Nothing.}
\author{Lars Relund \email{lars@relund.dk}}
\seealso{\code{\link{resetActions}}, \code{\link{fixAction}}.}
\arguments{\item{mdp}{The MDP loaded using \link{loadMDP}.}}
