% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run full deforestation prioritization pipeline}
\usage{
run_pipeline(
  base_path = NULL,
  folder_id,
  folder_id_tabelas,
  ano_inicial,
  ano_final,
  mes_inicial,
  mes_final,
  run_model = TRUE,
  run_viz = TRUE
)
}
\arguments{
\item{base_path}{Base folder where deter_data is located or will be created}

\item{folder_id}{Google Drive shapefile folder ID}

\item{folder_id_tabelas}{Google Drive tables folder ID}

\item{ano_inicial}{Initial year}

\item{ano_final}{Final year}

\item{mes_inicial}{Initial month}

\item{mes_final}{Final month}

\item{run_model}{Logical: run model step}

\item{run_viz}{Logical: run visualization step}
}
\description{
Run full deforestation prioritization pipeline
}
