% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp.R
\name{sample_plot}
\alias{sample_plot}
\title{Plot sample scores
Plots the sample scores data.frame for a given geneset. Data.frame must have Score, Sample and Class columns}
\usage{
sample_plot(sample_data, filename = file_name, directory = "", title = "",
  print = T, save = T, control_lab)
}
\arguments{
\item{sample_data}{Sample score information for a geneset. Must have columns "Sample", "Score" and "Class"}

\item{filename}{(optional) Name that will be added to the saved pdf filename}

\item{directory}{(optional) directory to save file}

\item{title}{(optional) Title name for graph}

\item{print}{(default T) Print plot to screen}

\item{save}{(default T) Save as a pdf file}

\item{control_lab}{(optional) Specifying control_lab will set the control class as light blue as a default}
}
\description{
Plot sample scores
Plots the sample scores data.frame for a given geneset. Data.frame must have Score, Sample and Class columns
}
