\name{setPolicyAction}
\alias{setPolicyAction}
\title{Set the action of a state to be in the current policy.}
\usage{
  setPolicyAction(mdp, sId, iA)
}
\arguments{
  \item{mdp}{The MDP loaded using \link{loadMDP}.}

  \item{sId}{The state id of the state.}

  \item{iA}{The action index of the state.}
}
\value{
  Nothing.
}
\description{
  Set the action of a state to be in the current policy.
}
\author{
  Lars Relund \email{lars@relund.dk}
}

