\name{setStateWeight}
\alias{setStateWeight}
\title{Set the weight of a state.}
\usage{
  setStateWeight(mdp, w, sId, wLbl)
}
\arguments{
  \item{mdp}{The MDP loaded using \link{loadMDP}.}

  \item{w}{The weight.}

  \item{sId}{The state id of the state.}

  \item{wLbl}{The label of the weight we consider.}
}
\value{
  Nothing.
}
\description{
  Set the weight of a state.
}
\author{
  Lars Relund \email{lars@relund.dk}
}

