\name{weightNames}
\alias{weightNames}
\title{Names of weights used in actions.}
\usage{weightNames(prefix, labels=actionWeightLbl.bin)}
\description{Names of weights used in actions.}
\author{Lars Relund \email{lars@relund.dk}}
\value{Vector of weight names.}
\arguments{\item{labels}{The HMDP binary file containing the weight labels.}}
