% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBinary2HMP.R
\name{weightNames}
\alias{weightNames}
\title{Names of weights used in actions.}
\usage{
weightNames(prefix = "", labels = "actionWeightLbl.bin")
}
\arguments{
\item{prefix}{A character string with the prefix added to the binary file names.}

\item{labels}{The HMDP binary file containing the weight labels.}
}
\value{
Vector of weight names.
}
\description{
Names of weights used in actions.
}
\author{
Lars Relund \email{lars@relund.dk}
}
